#Refactor GameScreen and MainMenu styles into separate files for better organization. #Added comments.
