+++ screens/GameScreen.js
@@ -6,6 +6,8 @@ import Physics from "../systems/Physics";
 import MoveSystem from "../systems/MoveSystem";
 import HoleSystem from "../systems/HoleSystem";
 import LevelData from "../levels/LevelData";
+import styles from "../styles/GameScreenStyles"; // Adjust the path as needed
+
 
 const GameScreen = () => {
     const [gameEngine, setGameEngine] = useState(null);
@@ -150,92 +152,6 @@ const GameScreen = () => {
     );
 };
 
-const styles = StyleSheet.create({
-    backgroundImage: {
-        flex: 1,
-        width: '100%',
-        height: '100%',
-    },
-    container: {
-        flex: 1,
-        backgroundColor: 'transparent',
-    },
-    leftSideHeaderContainer: {
-        position: 'absolute',
-        left: 0,
-        top: 20,
-        zIndex: 10,
-    },
-    levelBadge: {
-        backgroundColor: 'rgba(0,0,0,0.6)',
-        paddingVertical: 8,
-        paddingHorizontal: 16,
-        borderTopRightRadius: 15,
-        borderBottomRightRadius: 15,
-        shadowColor: '#000',
-        shadowOffset: { width: 2, height: 2 },
-        shadowOpacity: 0.3,
-        shadowRadius: 3,
-        elevation: 5,
-    },
-    levelText: {
-        fontSize: 20,
-        fontWeight: 'bold',
-        color: 'white',
-        textShadowColor: 'rgba(0, 0, 0, 0.75)',
-        textShadowOffset: { width: 1, height: 1 },
-        textShadowRadius: 2
-    },
-    gameContainer: {
-        flex: 1,
-    },
-    messageContainer: {
-        position: 'absolute',
-        top: '40%',
-        left: 0,
-        right: 0,
-        alignItems: 'center',
-        backgroundColor: 'rgba(0,0,0,0.7)',
-        padding: 20,
-    },
-    victoryText: {
-        color: 'gold',
-        fontSize: 38,
-        fontWeight: 'bold',
-        textAlign: 'center',
-        marginBottom: 20,
-    },
-    deathText: {
-        color: '#ff4444',
-        fontSize: 38,
-        fontWeight: 'bold',
-        textAlign: 'center',
-        marginBottom: 20,
-    },
-    nextLevelButton: {
-        backgroundColor: '#4CAF50',
-        paddingVertical: 12,
-        paddingHorizontal: 24,
-        borderRadius: 8,
-        marginTop: 10,
-    },
-    restartButton: {
-        backgroundColor: '#ff4444',
-        paddingVertical: 12,
-        paddingHorizontal: 24,
-        borderRadius: 8,
-        marginTop: 10,
-    },
-    nextLevelText: {
-        color: 'white',
-        fontSize: 18,
-        fontWeight: 'bold',
-    },
-    restartText: {
-        color: 'white',
-        fontSize: 18,
-        fontWeight: 'bold',
-    }
-});
+
 
 export default GameScreen;
